feat: :sparkles: Implemented custom NotFound, Error, and Loading pages
I created custom NotFound and Error pages to enhance user experience in case of missing pages or server errors.
The NotFound page displays a friendly message with an image and a link to return to the homepage, while the Error page provides a retry option when the connection to the server fails.
Additionally, I added a Loading page with a spinner animation to improve UX during data fetching. Global styles were updated to style these pages appropriately.