feat: :sparkles: Added standalone login page and improved API authentication handling
I created a dedicated login page that automatically redirects authenticated users to the homepage.
If the login modal is closed, the user is navigated back to the previous page for a smoother experience.
Additionally, I updated the Axios interceptors in apiConfig.js to redirect unauthorized users to the login page when accessing protected routes.
I also fixed a bug related to setting cookies when refreshing the authentication token, ensuring a more stable login session.